+++ my-app/src/core/interfaces/interfaces.ts
@@ -1,4 +1,4 @@
 export interface IUser {
 	name: string;
-	score: null | number;
+	score: number;
 }
