+++ my-app/src/pages/Title/title.module.scss
@@ -1,0 +1,11 @@
+@import "src/assets/scss/colors";
+@import "src/assets/scss/mixins";
+@import "src/assets/scss/animations";
+
+.title {
+  margin-top: 10px;
+
+  .digit {
+    color: $h1;
+  }
+}
