+++ my-app/src/store/usersCreateslice.ts
@@ -3,10 +3,11 @@ import { IUser } from '../core/interfaces/interfaces';
 import { fetchLeaders } from '../core/services/getRequest';
 
 interface InitialState {
+	day: number;
 	usersLeaders: IUser[];
 }
 
-const initialState: InitialState = { usersLeaders: [] };
+const initialState: InitialState = { day: 0, usersLeaders: [] };
 
 const users = createSlice({
 	name: 'users',
@@ -15,6 +16,8 @@ const users = createSlice({
 	extraReducers: builder => {
 		builder.addCase(fetchLeaders.fulfilled, (state, action: PayloadAction<IUser>) => {
 			state.usersLeaders.push(action.payload);
+			// eslint-disable-next-line no-param-reassign
+			state.day += 1;
 		});
 	},
 });
